Struct cargo_wasmer::Pack

#[non_exhaustive]
pub struct Pack {
    pub manifest: Manifest,
    pub workspace: Workspace,
    pub features: Features,
    pub debug: bool,
    pub out_dir: Option<PathBuf>,
}

Compile a Rust crate to a WAPM package.

Examples

use clap::Parser;
use cargo_wasmer::Pack;

let pack = Pack::parse();
let meta = pack.metadata()?;

for pkg in pack.resolve_packages(&meta) {
    let dest = pack.generate_wapm_package(pkg, meta.target_directory.as_ref())?;
    println!("Wrote the WAPM package for {} to {}", pkg.name, dest.display());
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

manifest: Manifestworkspace: Workspacefeatures: Featuresdebug: bool

Compile in debug mode.

out_dir: Option<PathBuf>

Where to save the compiled WAPM package(s)

Implementations

impl Pack

pub fn metadata(&self) -> Result<Metadata, Error>

Use the cargo metadata subcommand to learn about the workspace.

pub fn resolve_packages<'meta>( &self, metadata: &'meta Metadata ) -> Vec<&'meta Package>

Figure out which packages the user wants to pack.

pub fn generate_wasmer_package( &self, pkg: &Package, target_dir: &Path ) -> Result<PathBuf, Error>

Compile a crate to a Wasmer package and save it on disk.